AI-assistant/requirements.txt
pyttsx3
SpeechRecognition
pyaudio
playsound
wikipedia
python-weather
newsapi-python
openai
schedule
